refactor(studio): move shared preview utils into core/preview
