+++ packages/sanity/src/core/preview/index.ts
@@ -2,6 +2,8 @@ export * from './components/SanityDefaultPreview'
 export * from './components/SanityPreview'
 
 export {getPreviewPaths} from './utils/getPreviewPaths'
+export {getPreviewStateObservable} from './utils/getPreviewStateObservable'
+export {getPreviewValueWithFallback} from './utils/getPreviewValueWithFallback'
 export {prepareForPreview} from './utils/prepareForPreview'
 
 export * from './documentPreviewStore'
